+++ docs/demo/mandelbrot.js
@@ -21,14 +21,14 @@ const [width, height] = [dev.bufferWidth, dev.bufferHeight];
 
 const pingTexVal = Texture.create(dev, width, height, TextureInternalFormat.RG32F);
 const pingTexEsc = Texture.create(dev, width, height, TextureInternalFormat.R32F);
-const pingFbo = Framebuffer.withColor(dev, width, height, [
+const pingFbo = Framebuffer.create(dev, width, height, [
     pingTexVal,
     pingTexEsc,
 ]);
 
 const pongTexVal = Texture.create(dev, width, height, TextureInternalFormat.RG32F);
 const pongTexEsc = Texture.create(dev, width, height, TextureInternalFormat.R32F);
-const pongFbo = Framebuffer.withColor(dev, width, height, [
+const pongFbo = Framebuffer.create(dev, width, height, [
     pongTexVal,
     pongTexEsc,
 ]);